resolve signed/unsigned mismatch update throw/catch exceptions add in determineHeader for parser (done before Terence change to the new edit command style) update constants

+++ TimeWiseLogic/Task.cpp
@@ -39,7 +39,7 @@ Task::~Task(void){
 std::string Task::getDescription(){
 	return _taskDescription;
 }
-unsigned long long Task::getIndex(){
+int Task::getIndex(){
 	return _taskIndex;
 }
 
@@ -119,7 +119,7 @@ void Task::setCategory(std::string category){
 	_category = category;
 }
 
-void Task::setIndex(unsigned long long index){ 
+void Task::setIndex(int index){ 
 	assert(index>=0);
 	_taskIndex = index;
 }
